backend/requirements.txt
fastapi
uvicorn
python-multipart
pymongo
motor
beautifulsoup4
requests
opencv-python
easyocr
python-dotenv
